PiD decode: bf16 canvas/noise to claw back ~400MB at the DynamicVRAM edge
The 8K OOM is now comfy-aimdo's VRAM buffer failing to grow by 37MB -- we're at
the brink, not GB-short (pixel_dim=16 means activations are small). FLUX/TE are
aimdo 'dynamic' streamed models, not in current_loaded_models, so free_memory
can't evict them ('0 models unloaded'). Cheapest real headroom: the fp32
canvas+noise are ~810MB at 8K; bf16 halves that to ~400MB saved >> the 37MB
shortfall.

+++ dit360/nodes_pid.py
@@ -194,8 +194,10 @@ class DiT360PiDDecode:
         need = pid_model.model.memory_required([b * 2, 3, out_h, out_w])
         comfy.model_management.free_memory(need, dev)
 
-        # pixel-space canvas (ChromaRadiance: 3-ch, spatial = output pixels)
-        canvas = torch.zeros((b, 3, out_h, out_w), dtype=torch.float32)
+        # pixel-space canvas (ChromaRadiance: 3-ch, spatial = output pixels). bf16,
+        # not fp32: at 8K the canvas+noise alone are ~810MB in fp32 -- halving them
+        # is real headroom when the DynamicVRAM buffer is at its limit.
+        canvas = torch.zeros((b, 3, out_h, out_w), dtype=torch.bfloat16)
         noise = comfy.sample.prepare_noise(canvas, seed)
         _log(f"PiD Decode: freed VRAM for PiD; sampling on {dev} ...")
